Move GFG Arrays to Level 1 package
