+++ src/com/geeksForGeeks/arrays/level1/BalanceParans.java
@@ -1,4 +1,4 @@
-package com.geeksForGeeks.arrays;
+package com.geeksForGeeks.arrays.level1;
 
 import java.util.Scanner;
 import java.util.Stack;
