+++ src/com/geeksForGeeks/arrays/level1/ExtraIndex.java
@@ -1,4 +1,4 @@
-package com.geeksForGeeks.arrays;
+package com.geeksForGeeks.arrays.level1;
 
 import java.util.Scanner;
