+++ src/com/geeksForGeeks/arrays/level1/ReverseArr.java
@@ -1,4 +1,4 @@
-package com.geeksForGeeks.arrays;
+package com.geeksForGeeks.arrays.level1;
 
 import java.io.BufferedReader;
 import java.io.IOException;
