+++ src/com/geeksForGeeks/arrays/level1/SecondLargest.java
@@ -1,4 +1,4 @@
-package com.geeksForGeeks.arrays;
+package com.geeksForGeeks.arrays.level1;
 
 import java.util.Arrays;
 import java.util.List;
